renamed:    src/APNACOLLAGE/FirstClass.java -> src/APNACOLLAGE/Lecture_1/FirstClass.java 	renamed:    src/APNACOLLAGE/Stars1.java -> src/APNACOLLAGE/Lecture_1/Stars1.java 	renamed:    src/APNACOLLAGE/Input.java -> src/APNACOLLAGE/Lecture_2/Input.java 	renamed:    src/APNACOLLAGE/Var.java -> src/APNACOLLAGE/Lecture_2/Var.java
